USE THIS VERSION
Note:
when i tried to commit a change i got a conflicting error.
and my save became corrupted.
so i made a new one and put all of our changes in by hand. I even fixed
the bug u told me about.

REMEMBER: only commit to this version.
p.s: we may need to work out separate days to work on this so we don't
have this again.